# src/main/java/co/edu/unicauca/matricula_financiera/infrastructure/in/rest/student/dtoResponse/DocenteResponse.java
package co.edu.unicauca.matricula_financiera.infrastructure.in.rest.student.dtoResponse;

import lombok.AllArgsConstructor;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;

@Getter
@Setter
@NoArgsConstructor
@AllArgsConstructor
public class DocenteResponse {
    private String nombre;
    private String apellido;
}